🚚 Move typed-arrays arbitraries into /arbitrary
