+++ src/arbitrary/uint8Array.ts
@@ -1,0 +1,16 @@
+import { Arbitrary } from '../check/arbitrary/definition/Arbitrary';
+import { integer } from './integer';
+import {
+  IntArrayConstraints,
+  typedIntArrayArbitraryArbitraryBuilder,
+} from './_internals/builders/TypedIntArrayArbitraryBuilder';
+
+/**
+ * For Uint8Array
+ * @remarks Since 2.9.0
+ * @public
+ */
+export function uint8Array(constraints: IntArrayConstraints = {}): Arbitrary<Uint8Array> {
+  return typedIntArrayArbitraryArbitraryBuilder<Uint8Array, number>(constraints, 0, 255, Uint8Array, integer);
+}
+export { IntArrayConstraints };
